+++ internal/handler/response/response.go
@@ -11,17 +11,19 @@ import (
 type ErrMessage string
 
 const (
+	ErrorStatusExample ErrMessage = "some_error_status"
+
 	InternalServerError  ErrMessage = "internal_server_error"
-	TimedOut                        = "timed_out"
-	InvalidRequestParams            = "invalid_request_params"
-	ParamsMissing                   = "params_missing"
-	UnsupportedFileType             = "unsupported_file_type"
-	InvalidCredentials              = "invalid_credentials"
-	UserAlreadyExists               = "user_already_exists"
-	AuthHeaderEmpty                 = "auth_header_empty"
-	AuthHeaderInvalid               = "auth_header_invalid"
-	AuthTokenInvalid                = "auth_token_invalid"
-	Unauthorized                    = "unauthorized"
+	TimedOut             ErrMessage = "timed_out"
+	InvalidRequestParams ErrMessage = "invalid_request_params"
+	ParamsMissing        ErrMessage = "params_missing"
+	UnsupportedFileType  ErrMessage = "unsupported_file_type"
+	InvalidCredentials   ErrMessage = "invalid_credentials"
+	UserAlreadyExists    ErrMessage = "user_already_exists"
+	AuthHeaderEmpty      ErrMessage = "auth_header_empty"
+	AuthHeaderInvalid    ErrMessage = "auth_header_invalid"
+	AuthTokenInvalid     ErrMessage = "auth_token_invalid"
+	Unauthorized         ErrMessage = "unauthorized"
 )
 
 type Error struct {
@@ -55,20 +57,4 @@ func HandleError(c *gin.Context, err error) bool {
 		return true
 	}
 	return false
-}
-
-func MustGetUUID(c *gin.Context, key string) (string, bool) {
-	val, exists := c.Get(key)
-	if !exists {
-		NewErr(c, http.StatusUnauthorized, Unauthorized, nil, "Try logging in again.")
-		return "", false
-	}
-
-	uuid, ok := val.(string)
-	if !ok {
-		NewErr(c, http.StatusInternalServerError, InternalServerError, nil, "Unexpected error occurred.")
-		return "", false
-	}
-
-	return uuid, true
 }
