+++ internal/handler/v1/photos/photo.go
@@ -22,6 +22,18 @@ const (
 	FormPhotoBatchFiles = "batch_photo_files"
 )
 
+// @Summary Upload photo
+// @Description Upload single photo
+// @Tags photos
+// @Accept multipart/form-data
+// @Produce json
+// @Security JWTAuth
+// @Param photo_file formData file true "Photo file"
+// @Success 200 {object} response.UploadPhotoResponse
+// @Failure 400 {object} response.Error "Bad Request."
+// @Failure 401 {object} response.Error "Unauthorized."
+// @Failure 500 {object} response.Error "Unexpected error occurred."
+// @Router /api/v1/photos/upload [post]
 func (h *handler) uploadPhoto(c *gin.Context) {
 	ctx, cancel := context.WithTimeout(c, config.DefaultContextTimeout)
 	defer cancel()
@@ -48,12 +60,22 @@ func (h *handler) uploadPhoto(c *gin.Context) {
 		return
 	}
 
-	c.JSON(200, gin.H{
-		"status": "ok",
-		"id":     photoID,
-	})
+	response.NewOk(c, response.UploadPhotoResponse{PhotoID: photoID})
 }
 
+// @Summary Upload batch photos
+// @Description Upload multiple photos
+// @Tags photos
+// @Accept multipart/form-data
+// @Produce json
+// @Security JWTAuth
+// @Param batch_photo_files formData file true "Batch photo files"
+// @Success 200 {object} response.UploadBatchPhotosResponse
+// @Failure 206 {object} response.UploadBatchPhotosResponse
+// @Failure 400 {object} response.Error "Bad Request."
+// @Failure 401 {object} response.Error "Unauthorized."
+// @Failure 500 {object} response.Error "Unexpected error occurred."
+// @Router /api/v1/photos/upload/batch [post]
 func (h *handler) uploadBatchPhotos(c *gin.Context) {
 	ctx, cancel := context.WithTimeout(c, config.DefaultContextTimeout)
 	defer cancel()
@@ -100,10 +122,12 @@ func (h *handler) uploadBatchPhotos(c *gin.Context) {
 	c.JSON(respStatus, body)
 }
 
+// TODO: documetation
 func (h *handler) getPhotoVersions(c *gin.Context) {
 	ctx, cancel := context.WithTimeout(c, config.DefaultContextTimeout)
 	defer cancel()
 
+	// TODO: сделать проверку на права доступа к фото
 	idParam := c.Param("id")
 	id, err := strconv.Atoi(idParam)
 	if err != nil {
